Fix wait_for_server on playground installs
Playground installs don't create a package update task on the server, they
create a restart task instead, which has a different API call. This PR passes
around if the install is a playground or not and waits using the proper API
call.

This was hidden because the `package_update_status` returns a 200 code even
if the requested task isn't present (maybe that should be a 404? out of scope
here though). So now even if the server returns a 200 code for a particular
task id, we also check the status for `unknown` to see if the task id exists
on the server.